requirements.txt
imutils==0.5.1
keras==2.2.2
librosa==0.6.2
line-bot-sdk==1.8.0
numpy==1.14.5
PyAudio==0.2.11
scikit-learn==0.19.2
scipy==1.1.0
tqdm==4.25.0
urllib3==1.13.1
